🎨📝 Added build constraints to switch between json-iterator and std
